agentic_architecture_practice/pyproject.toml
[project]
name = "agentic-architecture"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ddgs>=9.11.2",
    "duckduckgo-search>=8.1.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-nebius>=0.1.3",
    "langchain-ollama>=1.0.1",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.10",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
]
